+++ libs/libcommon/include/common/getThreadId.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <cstdint>
 
 /// Obtain thread id from OS. The value is cached in thread local variable.
-unsigned getThreadId();
+uint64_t getThreadId();
